Added different board rooms. (See below)
This update allows for multiple board rooms. Only if two or more clients are connected to the same board, will
they see the same thing. Clients can connect to the same board by navigating to /boards/:id and giving the same id.

If, for example, two clients are connected to /boards/1234 and /boards/abcd, they will not see the same board.